refactor: 에외 클래스 패키지 수정
